+++ lib/xnix/compatibility/dbus_runtime_client.rb
@@ -57,6 +57,10 @@ module Xnix
         parse_dictionary(call("GetRunPlan", application_id))
       end
 
+      def backend_binding(application_id)
+        parse_dictionary(call("GetBackendBinding", application_id))
+      end
+
       def repair_plan(application_id, issue)
         parse_dictionary(call("GetRepairPlan", application_id, issue))
       end