+++ test/test_runtime_contract.rb
@@ -26,6 +26,7 @@ plasmoid = project_root.join("kde/plasmoids/org.xnix.compatibilitycenter/metadat
   GetDiagnostics
   GetEngineCatalog
   GetRunPlan
+  GetBackendBinding
   GetRepairPlan
   GetTestPlan
   GetTestResult